Fix routing issues
