--- frontend/src/components/auth/PrivateRoute.tsx
@@ -1,25 +1,0 @@
-import React, { useContext } from "react";
-import { Navigate, Route } from "react-router-dom";
-
-import { LOGIN_PAGE } from "../../constants/Routes";
-import AuthContext from "../../contexts/AuthContext";
-
-type PrivateRouteProps = {
-  element: React.ReactNode;
-  path: string;
-};
-
-const PrivateRoute: React.FC<PrivateRouteProps> = ({
-  element,
-  path,
-}: PrivateRouteProps) => {
-  const { authenticatedUser } = useContext(AuthContext);
-
-  return authenticatedUser ? (
-    <Route path={path} element={element} />
-  ) : (
-    <Navigate replace to={LOGIN_PAGE} />
-  );
-};
-
-export default PrivateRoute;
