+++ frontend/src/utils/AuthUtils.ts
@@ -5,9 +5,11 @@ export const decodeJWT = (token: string | null): jwt.JWTPayload | null => {
   return jwt.decodeJwt(token);
 };
 
-export const isUnexpiredToken = (token: string | null) => {
+export const shouldRenewToken = (token: string | null) => {
   const decodedToken = decodeJWT(token);
   if (!decodedToken) return false;
+
   if (decodedToken?.exp == null) return false;
+
   return decodedToken.exp <= Math.round(new Date().getTime() / 1000);
 };
